added more facet tests and readme fixed
