+++ README.md
@@ -2,14 +2,12 @@
 
 ## Project Description
 
-
-
-- This is a simple example implementation for EIP-2535 Diamond Standard. To learn about other implementations go here: https://github.com/mudgen/diamond
+- This is a simple example implementation for EIP-2535 Diamond Standard. To learn about other implementations go here: <https://github.com/mudgen/diamond>
 
 - The standard loupe functions have been gas-optimized in this implementation and can be called in on-chain transactions. However keep in mind that a diamond can have any number of functions and facets so it is still possible to get out-of-gas errors when calling loupe functions. Except for the facetAddress loupe function which has a fixed gas cost.
 
 - The DApp (Decentralised Application) was designed with the purpose of optimising DeFi (Decentralised Finance) users yields as they interact with the other DApps in the DeFi space.
-- Reference -(https://autofarm.gitbook.io/autofarm-network/)
+- Reference -<https://autofarm.gitbook.io/autofarm-network/>
 
 ---
 
@@ -21,8 +19,9 @@
 - [chai](https://www.chaijs.com/) for testing
 - [solidity-coverage](https://github.com/sc-forks/solidity-coverage) for generating code coverage
 - [truffle dashboard](https://trufflesuite.com/docs/truffle/how-to/use-the-truffle-dashboard/) for deplyment of the smart contracts
-  
+
 ---
+
 ## A typical top-level directory layout
 
 ```shell
@@ -38,6 +37,7 @@
 ├── README.md
 
 ```
+
 ---
 
 ## Instructions and hardhat commands
@@ -45,16 +45,19 @@
 ### Installation
 
 1- Clone this repo:
+
 ```shell
 git clone https://github.com/Shivansh070707/AutoFarm-Diamond-Standard
 ```
+
 2- Install the autofarm project dependencies. \
-  This will install the packages mentioned inside the `package.json` file.:
+ This will install the packages mentioned inside the `package.json` file.:
 
 ```shell
 cd AutoFarm-Diamond-Standard
 npm install or yarn install
 ```
+
 ### Deployment
 
 - Create the `.env` file following the `.env.example` file and fill in the contents. The values for the `ALCHEMY_API_KEY` can be obtained from the [Alchemy](https://www.alchemy.com/) website. If you don't have an account, you can create one for free.
@@ -82,6 +85,20 @@ npm install or yarn install
 
   ```shell
   npx hardhat test --network localhost
+  ```
+
+- Test Single files with typecheck:
+
+  ```shell
+  npx hardhat test ./test/autofarm.test.ts --network localhost --typecheck
+  ```
+
+  ```shell
+  npx hardhat test ./test/facet.test.ts --network localhost --typecheck
+  ```
+
+  ```shell
+  npx hardhat test ./test/stratx2.test.ts --network localhost --typecheck
   ```
 
 - Generate the code coverage report. \
@@ -121,36 +138,33 @@ npm install or yarn install
       ```
 
     - The browser will open up and then you have to connect with the MetaMask extension. Select the preferred network and the account to deploy the smart contract.
-   #### Deploy Autofarm
 
-    - Deploy the autofarm project smart contracts using either of the two ways mentioned above.
+#### Deploy Autofarm
 
-       ```shell
-       npx hardhat run .\scripts\autofarm.deploy.ts --network <NETWORK>  param1 param2
-       ```
+- for deploying autofarm diamond you need to modify the variables in './scripts/autofarm.deploy.ts'
 
-    - param1 is autov2 address \
-        example:<'0x20F5f006a0184883068bBF58fb0c526A8EEa8BFD'>
-    - param2 is farm name \
-        example: -<'AutoFarm Name>
+```shell
+  npx hardhat run .\scripts\autofarm.deploy.ts --network <NETWORK>
+```
 
-    ####  Deploy StratX2
-     - for deploying stratx2 diamond you need to modify the variables in './scripts/stratx2.pcs.ts'
-     ```shell
-      npx hardhat run .\scripts\stratx2.pcs.ts --network <NETWORK>
-     ```
-  
+#### Deploy StratX2
+
+- for deploying stratx2 diamond you need to modify the variables in './scripts/stratx2.deploy.ts'
+
+  ```shell
+   npx hardhat run .\scripts\stratx2.deploy.ts --network <NETWORK>
+  ```
 
 - `<NETWORK>` can be `localhost`, `goerli`, `plsTest`, `plsMain`, `mainnet` or any other network that is supported and defined in the `hardhat.config.ts`.
 
 - If the deployment uses the truffle dashboard method, then, switch to the browser and sign the deployment transactions from the MetaMask extension. For the deployment of the `initFunction.ts`, there will be **multiple transactions** that need to be signed via Metamask.
 
-- Unless `Done` is being shown in the terminal, the deployment process of the strategies are not over.
-
 - Otherwise, if the deployment uses the private key, the transactions will be signed automatically and will be shown in the terminal.
 
-- After the succesful deployment of the smart contracts, a `build` folder comprising the addresses and the ABIs of the deployed smart contracts will be generated.
+- After the successful deployment of the smart contracts, a `build` folder comprising the addresses and the ABIs of the deployed smart contracts will be generated.
+
 ---
+
 ## Smart contracts size
 
 - To generate the sizes of the smart contracts.
@@ -160,7 +174,9 @@ npm install or yarn install
   ```
 
   ![size](./smart-contract-size.png)
+
 ---
+
 ## Code coverage
 
 - To generate the code coverage report for the test cases of the smart contracts.
@@ -170,8 +186,9 @@ npm install or yarn install
   ```
 
   ![coverage](./code-coverage.png)
-  
+
 ---
+
 ## Facet Information
 
 - The contracts/diamond/Diamond.sol file shows an example of implementing a diamond.
@@ -185,14 +202,15 @@ npm install or yarn install
 - The scripts/autofarm.deploy.ts file shows how to deploy a Autofarm diamond.
 
 - The scripts/stratx2.ts & scripts/stratx2.pcs.ts file shows how to deploy a strat diamond.
-  
 - The test/facet.test.ts file gives tests for the diamondCut function and the Diamond Loupe functions.
-  
 - The test/autofarm.test.ts file gives tests for the autofarm facets.
 
 - The test/stratx2.test.ts file gives tests for the stratx2 facets.
+
 ---
+
 ## How to Get Started Making Your Diamond
+
 1- Reading and understand [EIP-2535 Diamonds](https://github.com/ethereum/EIPs/issues/2535). If something is unclear let me know!
 
 2- Use a diamond reference implementation. You are at the right place because this is the README for a diamond reference implementation.
@@ -210,14 +228,19 @@ The [LibDiamond.sol](https://github.com/Shivansh070707/AutoFarm-Diamond-Standard
 In order to call a function that exists in a diamond you need to use the ABI information of the facet that has the function.
 
 Here is an example that uses ethers.js:
+
 ```javascript
 let myUsefulFacet = await ethers.getContract(facetName,diamondDeployedAddress) as facetNameType;
 ```
+
 ```javascript
 string result = MyUsefulFacet(address(diamondContract)).getResult()
 ```
+
 ---
+
 ### Useful Links
+
 1.[Introduction to the Diamond Standard, EIP-2535 Diamonds](https://eip2535diamonds.substack.com/p/introduction-to-the-diamond-standard) \
 2.[EIP-2535 Diamonds](https://github.com/ethereum/EIPs/issues/2535) \
 3.[Understanding Diamonds on Ethereum](https://dev.to/mudgen/understanding-diamonds-on-ethereum-1fb) \
@@ -227,9 +250,13 @@ string result = MyUsefulFacet(address(diamondContract)).getResult()
 7.[buidler-deploy supports diamonds](https://github.com/wighawag/buidler-deploy/)
 
 ---
+
 ## Author
+
 Shivansh Shrivastava
 
 ---
+
 ### License
+
 MIT license. See the license file. Anyone can use or modify this software for their purposes.
