+++ hardhat.config.ts
@@ -39,11 +39,11 @@ const config: HardhatUserConfig = {
     'truffle-dashboard': {
       url: 'http://localhost:24012/rpc',
     },
-    // goerli: {
-    //   url: `https://eth-goerli.g.alchemy.com/v2/${ALCHEMY_API_KEY}`,
-    //   accounts: [`0x${PRIVATE_KEY}`],
-    //   chainId: 5,
-    // },
+    goerli: {
+      url: `https://eth-goerli.g.alchemy.com/v2/${ALCHEMY_API_KEY}`,
+      accounts: [`0x${PRIVATE_KEY}`],
+      chainId: 5,
+    },
   },
   etherscan: {
     apiKey: GOERLI_API_KEY,
